requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
pillow==10.1.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
apscheduler==3.10.4
openai==1.3.0
pytz==2023.3
replicate==0.20.0
